% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctylevel}
\alias{ctylevel}
\title{County level census data}
\format{A data.table with 3135 rows and 164 variables}
\usage{
ctylevel
}
\description{
Description of the columns exists in an Excel file.
}
\details{
\describe{
\item{marketkey}{Unique market id. Should really be replaced wby marketid}
\item{ppc}{ppc value for the market}

}
}
\author{
Rahul Dodhia
}
\keyword{datasets}

