% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_potential.R
\name{traffic_potential}
\alias{traffic_potential}
\title{Traffic Potential Tool}
\usage{
traffic_potential(...)
}
\description{
A wrapper for shiny::runApp(). Does not take any arguments.
}
\author{
Rahul Dodhia
}

